Refactor(ble): SPI 수신 경로 uint8_t 전환 (B2)
B1 에서는 호출부를 건드리지 않으려고 Rx_DataPacket 을 int 로 남겨 뒀다.
이번에는 수신 경로 전체를 바이트 타입으로 정합시킨다. 프로토콜이 바이트
단위이므로 패킷을 int 배열로 다룰 이유가 없다.

변경
- tdc_hal_spi.c/.h: Rx_DataPacket 을 uint8_t 로,
  tdc_hal_spi_get_rx_packet_addr() 반환형을 uint8_t* 로
- tdc_ble_communication.c: p_Rx_dataPacket 및 fetch_readDataForBleSetting()
- tdc_ble_remote.c/.h, tdc_ble_mapping.c/.h: fetch_packet() 시그니처
- tdc_dfu_ble_boot.c/.h, tdc_dfu_ble_ota.c/.h: fetch 진입점 3개 +
  내부 static 함수 7개 (_fetch_packet_boot_info/_boot_select/_command/_data,
  _handle_command_option_write/_read/_size)

연쇄 전환은 컴파일러가 3단계에 걸쳐 전부 잡아냈다. 포인터 타입 불일치는
경고로 드러나므로 누락이 생길 수 없는 방식이다.

값 손실 없음
패킷에서 값을 꺼내 int 구조체 필드에 넣는 코드는 전부 uint8_t -> int
확대 변환이라 안전하다. 16비트 조립(packet[i] << 8 | packet[i+1])도
정수 승격 규칙상 int 로 계산되므로 결과가 같다.

검증
- CLI 빌드 에러 0 / 경고 0
- bss 21864 -> 21804 (Rx_DataPacket 80 -> 20 바이트)
- 실기 확인 대기

+++ src/2__cm3/source/ble/tdc_ble_communication.c
@@ -26,7 +26,7 @@ typedef struct
 
 BleSettingPacket bleSettingPacket;
 
-void fetch_readDataForBleSetting(const int *Rx_dataPacket)
+void fetch_readDataForBleSetting(const uint8_t *Rx_dataPacket)
 {
     bleSettingPacket.command = Rx_dataPacket[0];
 
@@ -231,8 +231,8 @@ void setting_nrf_ble_adv_info(void)
 
 ST__BLE_COMMUNICATION_STATE tdc_ble_communication_step(ST__ISD_STATUS isd_state)
 {
-    int         i;
-    static int *p_Rx_dataPacket;
+    int             i;
+    static uint8_t *p_Rx_dataPacket;
 
     tdc_hal_spi_comm_state_t         communicationState;
     ST__MAPPING_STATE           mappingState;
